Extract device template functions into a devices Jinja2 extension
