+++ homeassistant/helpers/template/extensions/__init__.py
@@ -3,6 +3,7 @@
 from .base64 import Base64Extension
 from .collection import CollectionExtension
 from .crypto import CryptoExtension
+from .devices import DeviceExtension
 from .floors import FloorExtension
 from .labels import LabelExtension
 from .math import MathExtension
@@ -13,6 +14,7 @@ __all__ = [
     "Base64Extension",
     "CollectionExtension",
     "CryptoExtension",
+    "DeviceExtension",
     "FloorExtension",
     "LabelExtension",
     "MathExtension",
